created a mini benchmark to determine the best sat solver for CAGP

+++ src/CAGPSolverSAT.py
@@ -1,23 +1,25 @@
+import time
 from pysat.solvers import Solver
 import rustworkx as rx
 from pyvispoly import PolygonWithHoles
+from threading import Timer
 
 class CAGPSolverSAT:
     
-    def __init__(self, K: int, guard_to_witnesses: dict[int, set[int]], witness_to_guards: dict[int, set[int]], initial_witnesses: list[int], all_witnesses: set[int], GC: rx.PyGraph, guard_color_constraints: bool, solution: list[list[int]]=None) -> list[tuple[int, int]]:
+    def __init__(self, K: int, guard_to_witnesses: dict[int, set[int]], witness_to_guards: dict[int, set[int]], initial_witnesses: list[int], all_witnesses: set[int], GC: rx.PyGraph, guard_color_constraints: bool=False, solution: list[list[int]]=None, solver_name="Gluecard4") -> list[tuple[int, int]]:
         self.GC = GC
         self.K = K
         self.guard_to_witnesses = guard_to_witnesses
         self.witness_to_guards = witness_to_guards
         self.witnesses = initial_witnesses
+        self.number_of_witnesses = len(initial_witnesses)
         self.all_witnesses = all_witnesses
-        self.solver = Solver(name='Gluecard4', with_proof=False)
+        self.solver = Solver(name=solver_name, with_proof=False)
         self.__make_vars()
         self.__add_witness_covering_constraints()
         self.__add_conflicting_guards_constraints()
         if guard_color_constraints:
             self.__add_guard_coloring_constraints()
-        # self.__add_edge_clique_cover_constraints()
 
     def __make_vars(self):
         self.guard_to_var = dict()
@@ -72,27 +74,47 @@ class CAGPSolverSAT:
         return missing_witnesses
 
     def __solve(self, assumptions=None):
-        self.solver.solve(assumptions=assumptions)
-        if not self.solver.solve(assumptions=assumptions):
+        passed_time = time.time() - self.start
+        if passed_time > self.time_limit:
+            self.interrupt()
+        print(self.time_limit - passed_time, 'seconds left')
+        timer = Timer(interval=(self.time_limit - passed_time), function=self.interrupt)
+        timer.start()
+
+        solution = self.solver.solve_limited(assumptions=assumptions, expect_interrupt=True)
+        if not solution:
             print('No solution found')
-            return None
+            timer.cancel()
+            return []
         print('Solution found')
+
+        timer.cancel()
         return [self.var_to_guard[var] for var in self.solver.get_model() if var > 0]
     
     def solve(self, color_lim: int=0):
+        self.start = time.time()
+        self.time_limit = 600
+        self.timeout = False
+
         if color_lim == 0:
             print('Starting binary search')
             color_lim, solution = self.binary_search()
         else:
-            solution = self.linear_search(color_lim)
+            color_lim, solution = self.linear_search(color_lim)
+
+        if self.timeout:
+            print('Timeout')
+            return color_lim, solution, iteration, self.number_of_witnesses, "timeout"
 
         print('Checking coverage')
         missing_witnesses = self.__check_coverage(solution)
+        print('Number of missing witnesses:', len(missing_witnesses))
+        self.number_of_witnesses += len(missing_witnesses)
 
-        iteration = 0
-        while(missing_witnesses):
+        iteration = 1
+        while missing_witnesses:
             iteration += 1
-            print('Adding new witnesses for missing area ({iteration})')
+            print(f'Adding new witnesses for missing area ({iteration})')
 
             for witness in missing_witnesses:
                 subset = []
@@ -103,15 +125,26 @@ class CAGPSolverSAT:
 
             print('Starting linear ascent')
             color_lim, solution = self.linear_ascent(color_lim)
+
+            if self.timeout:
+                print('Timeout')
+                return color_lim, solution, iteration, self.number_of_witnesses, "timeout"
+
             print('Checking coverage')
             missing_witnesses = self.__check_coverage(solution)
-        return solution
+            self.number_of_witnesses += len(missing_witnesses)
+
+        # Process solution such that each guard is only assigned one color
+        seen = dict()
+        solution = [seen.setdefault(g[0], g) for g in solution if g[0] not in seen]
+
+        return color_lim, solution, iteration, self.number_of_witnesses, "success"
 
     def binary_search(self):
         lower = 1
         upper = self.K
         optimal_solution = None
-        while lower < upper:
+        while lower < upper and not self.timeout:
             mid = (lower + upper) // 2
             print(f'Checking for {mid} colors')
             solution = self.__solve(assumptions=self.__deactivate_guards(mid))
@@ -127,7 +160,7 @@ class CAGPSolverSAT:
     
     def linear_ascent(self, color_lim: int):
         solution = self.__solve(assumptions=self.__deactivate_guards(color_lim))
-        while(not solution and color_lim <= self.K):
+        while not solution and color_lim <= self.K and not self.timeout:
             color_lim += 1
             print(f'Checking for {color_lim} colors')
             solution = self.__solve(assumptions=self.__deactivate_guards(color_lim))
@@ -137,16 +170,20 @@ class CAGPSolverSAT:
         print(f'Checking for {color_lim} colors')
         solution = self.__solve(assumptions=self.__deactivate_guards(color_lim))
         if solution:
-            while(solution and color_lim > 0):
+            while solution and color_lim > 0 and not self.timeout:
                 color_lim -= 1
                 print(f'Checking for {color_lim} colors')
                 solution = self.__solve(assumptions=self.__deactivate_guards(color_lim))
         else:
-            while(not solution and color_lim <= self.K):
+            while not solution and color_lim <= self.K and not self.timeout:
                 color_lim += 1
                 print(f'Checking for {color_lim} colors')
                 solution = self.__solve(assumptions=self.__deactivate_guards(color_lim))
-        return solution
+        return color_lim, solution
+    
+    def interrupt(self):
+        self.solver.interrupt()
+        self.timeout = True
 
     def __del__(self):
         """
@@ -155,4 +192,5 @@ class CAGPSolverSAT:
         There seem to occur resource leaks when you leave this out,
         so it should be sufficient to let the solvers clean up at the garbage collection.
         """
-        self.solver.delete()
+        if hasattr(self, 'solver') and self.solver is not None:
+            self.solver.delete()
